+++ src/client/engine/filesystem/server_namespace_provider.h
@@ -17,8 +17,7 @@ typedef struct coduomp_server_namespace_provider_s {
         const char *serverName, char resolvedServer[MAX_QPATH],
         char resolvedMod[FS_PACK_NAME_SIZE]);
     int32_t (*listCachedDemos)(const char *modName);
-    const char *(*stateRoot)(const char *ordinaryHomeRoot);
-    const char *(*contentRoot)(const char *ordinaryHomeRoot);
+    const char *(*root)(const char *ordinaryHomeRoot);
     qboolean (*allowsSearchpath)(const searchpath_t *searchpath);
     void (*promoteCurrentConfig)(void);
     void (*clearConfigs)(void);